patron repository

+++ resources/views/tickets/details.blade.php
@@ -27,20 +27,22 @@
                 <span class="label label-info">{{ $user->name }}</span>
             @endforeach    
             </p>
-            @if (! currentUser()->hasVoted($ticket))
-            {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
-                <button type="submit" class="btn btn-primary">
-                    <span class="glyphicon glyphicon-thumbs-up"></span> Votar
-                </button>
-            {!! Form::close() !!}
-            @else
-            {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
-                <button type="submit" class="btn-warning">
-                    <span class="glyphicon glyphicon-thumbs-down"></span> Quitar voto
-                </button>
-            {!! Form::close() !!}
+            @if(Auth::check())
+                @if (! currentUser()->hasVoted($ticket))
+                {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
+                    <button type="submit" class="btn btn-primary">
+                        <span class="glyphicon glyphicon-thumbs-up"></span> Votar
+                    </button>
+                {!! Form::close() !!}
+                @else
+                {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
+                    <button type="submit" class="btn-warning">
+                        <span class="glyphicon glyphicon-thumbs-down"></span> Quitar voto
+                    </button>
+                {!! Form::close() !!}
+                @endif
             @endif
-
+                
             <h3>Nuevo Comentario</h3>
 
             @include('partials/errors')
